Add DataManager which uploads to a server.
Key value pairs and files can be uploaded to the server.

All uploads are performed within a foreground service which
completes actions when data is not being collected.
Once the data is verified as successfully uploaded
it is deleted from device.

An account credential is needed with the server, which can
be created with the use of an admin password, after which
the credentials are stored on device and the app user
does not need to worry about entering any usernames or
passwords.

Still to be completed:
* Using the DataManager to save clip data
* Downloading directives (such as phrase lists) from the server.
* Hiding the account creation workflow from the typical user.
* Removing Photos based upload code.
* Comments and general cleanup.