Added server directives.
Phrases can be downloaded.
The app itself can be upgraded from the server (persuant to a lot of caveats
and the user clicking through some approvals).
The data manager uses global state for loginToken and phrases, so updating
the data in an activity will show it in the foreground service and vice versa.